requirements.txt
# API Automation Framework - Conduit API
# Python 3.8+

requests>=2.28.0
pytest>=7.0.0
python-dotenv>=1.0.0
pytest-html>=3.1.0
jsonschema>=4.17.0
